vector-db/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
faiss-cpu==1.7.4  # For vector similarity search
numpy==1.24.3
python-dotenv==1.0.0
pydantic==1.10.9
pickle5==0.0.12  # For Python 3.8+ compatibility
httpx==0.24.1  # For health checks